+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Python dependencies not being discovered when running pip install. Since dependencies
+  are fetched with Nix, we instead tell pip to not "verify" them when doing the
+  installation. This was a problem for example for PySide2 and Shiboken.
+
 ## [6.2.0] - 2022-07-11
 
 ### Added