+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [6.3.0] - 2022-09-19
+
 ### Fixed
 - Python dependencies not being discovered when running pip install. Since dependencies
   are fetched with Nix, we instead tell pip to not "verify" them when doing the